pull out the nelder-mead stuff and bring in heuristics

+++ src/config.rs
@@ -17,5 +17,5 @@ pub const LED_COUNT: usize = 32;
 pub const LED_REFRESH_INTERVAL: Duration = Duration::from_millis(100);
 
 pub const IDLE_TEMPERATURE: f32 = 60.0;
-pub const BOILER_POWER: u16 = 2000;
+pub const BOILER_POWER: f32 = 2000.0;
 pub const INITIAL_TEMPERATURE: f32 = 25.0;
